Added support to autogenerate qcas.bat files
New feature to generate qcas.bat files, this still needs to be checked by the user. 
Removed the prompt to open TSL files once complete and instead prompt to generate the qcas.bat file, if user says "Yes", then the script will prompt for the latest qcas.bat file which the script will use to generate the new bat file. 
qcas.bat file is not overwritten, a new file "qcas_generated.bat" is created in its place in the same directory as this script.